+++ lib/atoms/score.dart
