+++ lib/atoms/timer.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+class TimerState extends State<Timer> {
+  void build() {}
+}
+
+class Timer extends StatefulWidget {}
